+++ lib/class/xml_data.class.php
@@ -699,8 +699,8 @@ class XML_Data
                 "\t<name><![CDATA[" . $podcast->f_title . "]]></name>\n" .
                 "\t<description><![CDATA[" . $podcast->description . "]]></description>\n" .
                 "\t<language><![CDATA[" . $podcast->f_language . "]]></language>\n" .
-                "\t<copyright>" . $podcast->f_copyright . "</copyright>\n" .
-                "\t<generator>" . $podcast->f_generator . "</generator>\n" .
+                "\t<copyright><![CDATA[" . $podcast->f_copyright . "]]></copyright>\n" .
+                "\t<generator><![CDATA[" . $podcast->f_generator . "]]></generator>\n" .
                 "\t<website><![CDATA[" . $podcast->f_website . "]]></website>\n" .
                 "\t<build_date><![CDATA[" . $podcast->f_lastbuilddate . "]]></build_date>\n" .
                 "\t<sync_date><![CDATA[" . $podcast->f_lastsync . "]]></sync_date>\n" .